// web-ui/api/types.gen.ts
export type Beverage = {
    readonly id: number;
    /**
     * Name of the beverage, such as "Potrero Pale".
     */
    name: string;
    producer: BeverageProducer;
    beverage_type?: BeverageTypeEnum;
    /**
     * Beverage style within type, eg "Pale Ale", "Pinot Noir".
     */
    style?: string | null;
    /**
     * Free-form description of the beverage.
     */
    description?: string | null;
    picture: Picture;
    /**
     * Date of production, for wines or special/seasonal editions
     */
    vintage_year?: string | null;
    /**
     * ABV Percentage
     * Alcohol by volume, as percentage (0.0-100.0).
     */
    abv_percent?: number | null;
    /**
     * Calories per mL of beverage.
     */
    calories_per_ml?: number | null;
    /**
     * Carbohydrates per mL of beverage.
     */
    carbs_per_ml?: number | null;
    /**
     * Color (Hex Value)
     * Approximate beverage color
     */
    color_hex?: string;
    /**
     * Original gravity (beer only).
     */
    original_gravity?: number | null;
    /**
     * Final gravity (beer only).
     */
    specific_gravity?: number | null;
    /**
     * SRM Value
     * Standard Reference Method value (beer only).
     */
    srm?: number | null;
    /**
     * IBUs
     * International Bittering Units value (beer only).
     */
    ibu?: number | null;
    /**
     * Star rating for beverage (0: worst, 5: best)
     */
    star_rating?: number | null;
    /**
     * Untappd.com resource ID (beer only).
     */
    untappd_beer_id?: number | null;
};
// ...
export type Drink = {
    readonly id: number;
    /**
     * Flow sensor ticks, never changed once recorded.
     */
    readonly ticks: number;
    /**
     * Calculated (or set) Drink volume.
     */
    readonly volume_ml: number;
    /**
     * Date and time of pour.
     */
    readonly time: string;
    /**
     * Time in seconds taken to pour this Drink.
     */
    readonly duration: number;
    user: User;
    keg: Keg;
    /**
     * Session where this Drink is grouped.
     */
    readonly session_id: number | null;
    /**
     * Comment from the drinker at the time of the pour.
     */
    shout?: string | null;
    picture: Picture;
};

export type DrinkReassignRequest = {
    username: string;
};

export type DrinkingSession = {
    readonly id: number;
    start_time: string;
    end_time: string;
    volume_ml?: number;
    timezone?: string;
    name?: string | null;
    readonly stats: unknown;
};
// ...
export type Keg = {
    readonly id: number;
    beverage: Beverage;
    /**
     * Keg container type, used to initialize keg's full volume
     *
     * * `mini` - Mini Keg (5 L)
     * * `corny-2_5-gal` - Corny Keg (2.5 gal)
     * * `corny-3-gal` - Corny Keg (3.0 gal)
     * * `corny` - Corny Keg (5 gal)
     * * `sixth` - Sixth Barrel (5.17 gal)
     * * `euro-30-liter` - European DIN (30 L)
     * * `euro-half` - European Half Barrel (50 L)
     * * `quarter` - Quarter Barrel (7.75 gal)
     * * `euro` - European Full Barrel (100 L)
     * * `half-barrel` - Half Barrel (15.5 gal)
     * * `other` - Other
     */
    keg_type?: KegTypeEnum;
    /**
     * Computed served volume.
     */
    readonly served_volume_ml: number;
    /**
     * Full volume of this Keg; usually set automatically from keg_type.
     */
    full_volume_ml?: number;
    /**
     * Time the Keg was first tapped.
     */
    readonly start_time: string;
    /**
     * Time the Keg was finished or disconnected.
     */
    readonly end_time: string;
    /**
     * Current keg state.
     *
     * * `available` - Available
     * * `on_tap` - On tap
     * * `finished` - Finished
     */
    status: KegStatusEnum;
    /**
     * User-visible description of the Keg.
     */
    description?: string | null;
    /**
     * Amount of beverage poured without an associated Drink.
     */
    readonly spilled_ml: number;
    /**
     * Private notes about this keg, viewable only by admins.
     */
    notes?: string | null;
    readonly illustration: string;
    readonly illustration_thumbnail: string;
    readonly stats: unknown;
};
// ...
export type KegStatusEnum = 'available' | 'on_tap' | 'finished';
// ...
export type KegTypeEnum = 'mini' | 'corny-2_5-gal' | 'corny-3-gal' | 'corny' | 'sixth' | 'euro-30-liter' | 'euro-half' | 'quarter' | 'euro' | 'half-barrel' | 'other';
// ...
export type Picture = {
    readonly id: number;
    readonly resized_url: string;
    readonly resized_png_url: string;
    readonly thumbnail_url: string;
    readonly thumbnail_png_url: string;
    caption: string;
    /**
     * User that owns/uploaded this picture
     */
    readonly user_id: number | null;
    /**
     * Keg this picture was taken with, if any.
     */
    readonly keg_id: number | null;
    /**
     * Session this picture was taken with, if any.
     */
    readonly session_id: number | null;
};
// ...
export type SystemEvent = {
    readonly id: number;
    /**
     * Type of event.
     *
     * * `drink_poured` - Drink poured
     * * `session_started` - Session started
     * * `session_joined` - User joined session
     * * `keg_tapped` - Keg tapped
     * * `keg_volume_low` - Keg volume low
     * * `keg_ended` - Keg ended
     */
    kind: SystemEventKindEnum;
    /**
     * Time of the event.
     */
    time: string;
    drink: Drink;
    user: User;
    keg: Keg;
    session: DrinkingSession;
};

/**
 * * `drink_poured` - Drink poured
 * * `session_started` - Session started
 * * `session_joined` - User joined session
 * * `keg_tapped` - Keg tapped
 * * `keg_volume_low` - Keg volume low
 * * `keg_ended` - Keg ended
 */
export type SystemEventKindEnum = 'drink_poured' | 'session_started' | 'session_joined' | 'keg_tapped' | 'keg_volume_low' | 'keg_ended';
// ...
export type User = {
    readonly id: number;
    /**
     * Required. 30 characters or fewer. Letters, numbers and @/./+/-/_ characters
     */
    username: string;
    /**
     * Full name, will be shown in some places instead of username
     */
    display_name?: string;
    /**
     * Staff status
     * Designates whether the user can log into this admin site.
     */
    readonly is_staff: boolean;
    /**
     * Active
     * Designates whether this user should be treated as active. Unselect this instead of deleting accounts.
     */
    readonly is_active: boolean;
    picture: Picture;
};